patch: Move state modules to the same directory
