+++ packages/vest/src/core/produce/hasByGroup/index.js
@@ -1,4 +1,4 @@
-import useTestObjects from '../../test/useTestObjects';
+import useTestObjects from '../../state/useTestObjects';
 import { hasLogic } from '../has';
 
 /**
