+++ packages/vest/src/core/state/usePending/index.js
@@ -1,4 +1,4 @@
-import state from '../../../../state';
+import state from '..';
 
 export default state.registerHandler(() => ({
   pending: [],
