+++ packages/vest/src/core/state/useTestObjects/index.js
@@ -1,3 +1,3 @@
-import state from '../../state';
+import state from '..';
 
 export default state.registerHandler(() => []);
